+++ client/web/src/insights/pages/dashboards/components/dashboard-select/DashboardSelect.story.tsx
@@ -1,5 +1,5 @@
 import { storiesOf } from '@storybook/react'
-import React from 'react'
+import React, { useState } from 'react'
 
 import { WebStory } from '../../../../../components/WebStory'
 import { InsightDashboard, InsightsDashboardType } from '../../../../core/types'
@@ -30,6 +30,8 @@ const DASHBOARDS: InsightDashboard[] = [
         type: InsightsDashboardType.Personal,
         id: '102',
         title: 'Code Insights dashboard',
+        builtIn: false,
+        settingsKey: 'codeInsightsDasbhoard',
         insightIds: [],
         owner: {
             id: '101',
@@ -40,6 +42,8 @@ const DASHBOARDS: InsightDashboard[] = [
         type: InsightsDashboardType.Personal,
         id: '103',
         title: 'Experimental Insights dashboard',
+        builtIn: false,
+        settingsKey: 'experimentalInsightsDashboard',
         insightIds: [],
         owner: {
             id: '101',
@@ -61,6 +65,8 @@ const DASHBOARDS: InsightDashboard[] = [
         type: InsightsDashboardType.Organization,
         id: '105',
         title: 'Loooong looo0000ong name of dashboard',
+        builtIn: false,
+        settingsKey: 'looonglooongDashboard',
         insightIds: [],
         owner: {
             id: '104',
@@ -71,6 +77,8 @@ const DASHBOARDS: InsightDashboard[] = [
         type: InsightsDashboardType.Organization,
         id: '106',
         title: 'Loooong looo0000ong name of dashboard',
+        builtIn: false,
+        settingsKey: 'looonglooongDashboard',
         insightIds: [],
         owner: {
             id: '104',
@@ -79,4 +87,8 @@ const DASHBOARDS: InsightDashboard[] = [
     },
 ]
 
-add('DashboardSelect', () => <DashboardSelect dashboards={DASHBOARDS} />)
+add('DashboardSelect', () => {
+    const [value, setValue] = useState<string>()
+
+    return <DashboardSelect value={value} dashboards={DASHBOARDS} onSelect={dashboard => setValue(dashboard.id)} />
+})