+++ cmd/frontend/graphqlbackend/batches.go
@@ -228,6 +228,14 @@ type MergeChangesetsArgs struct {
 	Squash bool
 }
 
+type CreateBatchSpecExecutionArgs struct {
+	Spec string
+}
+
+type CloseChangesetsArgs struct {
+	BulkOperationBaseArgs
+}
+
 type BatchChangesResolver interface {
 	//
 	// MUTATIONS
@@ -258,6 +266,8 @@ type BatchChangesResolver interface {
 	CreateChangesetComments(ctx context.Context, args *CreateChangesetCommentsArgs) (BulkOperationResolver, error)
 	ReenqueueChangesets(ctx context.Context, args *ReenqueueChangesetsArgs) (BulkOperationResolver, error)
 	MergeChangesets(ctx context.Context, args *MergeChangesetsArgs) (BulkOperationResolver, error)
+	CreateBatchSpecExecution(ctx context.Context, args *CreateBatchSpecExecutionArgs) (BatchSpecExecutionResolver, error)
+	CloseChangesets(ctx context.Context, args *CloseChangesetsArgs) (BulkOperationResolver, error)
 
 	// Queries
 
@@ -691,4 +701,15 @@ type ChangesetCountsResolver interface {
 	OpenApproved() int32
 	OpenChangesRequested() int32
 	OpenPending() int32
+}
+
+type BatchSpecExecutionResolver interface {
+	ID() graphql.ID
+	InputSpec() string
+	State() string
+	StartedAt() *DateTime
+	FinishedAt() *DateTime
+	Failure() *string
+	PlaceInQueue() *int32
+	BatchSpec(ctx context.Context) (BatchSpecResolver, error)
 }