+++ enterprise/cmd/frontend/internal/auth/githuboauth/session.go
@@ -20,6 +20,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/auth/oauth"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	esauth "github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
 	githubsvc "github.com/sourcegraph/sourcegraph/internal/extsvc/github"
@@ -29,6 +30,7 @@ import (
 
 type sessionIssuerHelper struct {
 	*extsvc.CodeHost
+	db          dbutil.DB
 	clientID    string
 	allowSignup bool
 	allowOrgs   []string
@@ -125,12 +127,12 @@ func (s *sessionIssuerHelper) CreateCodeHostConnection(ctx context.Context, toke
 	}
 
 	// We have a special flow enabled when a user added code host has been created
-	// without 'repo` scope and we then enable private code on the instance. In this
+	// without `repo` scope and we then enable private code on the instance. In this
 	// case we allow the user to request the additional scope. This means that at
 	// this point we may already have a code host and we just need to update the
 	// token with the new one.
 
-	tx, err := database.GlobalExternalServices.Transact(ctx)
+	tx, err := database.ExternalServices(s.db).Transact(ctx)
 	if err != nil {
 		return
 	}