+++ enterprise/cmd/frontend/internal/auth/gitlaboauth/session.go
@@ -17,6 +17,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/auth/oauth"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gitlab"
 	"github.com/sourcegraph/sourcegraph/internal/jsonc"
@@ -26,6 +27,7 @@ import (
 type sessionIssuerHelper struct {
 	*extsvc.CodeHost
 	clientID string
+	db       dbutil.DB
 }
 
 func (s *sessionIssuerHelper) GetOrCreateUser(ctx context.Context, token *oauth2.Token, anonymousUserID, firstSourceURL string) (actr *actor.Actor, safeErrMsg string, err error) {
@@ -93,12 +95,12 @@ func (s *sessionIssuerHelper) CreateCodeHostConnection(ctx context.Context, toke
 	}
 
 	// We have a special flow enabled when a user added code host has been created
-	// without 'api` scope and we then enable private code on the instance. In this
+	// without `api` scope and we then enable private code on the instance. In this
 	// case we allow the user to request the additional scope. This means that at
 	// this point we may already have a code host and we just need to update the
 	// token with the new one.
 
-	tx, err := database.GlobalExternalServices.Transact(ctx)
+	tx, err := database.ExternalServices(s.db).Transact(ctx)
 	if err != nil {
 		return
 	}