+++ enterprise/internal/insights/resolvers/resolver.go
@@ -32,17 +32,18 @@ func New(timescale, postgres dbutil.DB) graphqlbackend.InsightsResolver {
 // clock for timestamps.
 func newWithClock(timescale, postgres dbutil.DB, clock func() time.Time) *Resolver {
 	return &Resolver{
-		insightsStore:   store.NewWithClock(timescale, clock),
+		insightsStore:   store.NewWithClock(timescale, store.NewInsightPermissionStore(postgres), clock),
 		workerBaseStore: basestore.NewWithDB(postgres, sql.TxOptions{}),
 		settingStore:    database.Settings(postgres),
 	}
 }
 
 func (r *Resolver) Insights(ctx context.Context, args *graphqlbackend.InsightsArgs) (graphqlbackend.InsightConnectionResolver, error) {
-	idList := make([]string, 0)
+	var idList []string
 	if args != nil && args.Ids != nil {
-		for _, id := range *args.Ids {
-			idList = append(idList, string(id))
+		idList = make([]string, len(*args.Ids))
+		for i, id := range *args.Ids {
+			idList[i] = string(id)
 		}
 	}
 	return &insightConnectionResolver{